requirements.txt
fastapi
uvicorn[standard]
feedparser
python-dateutil

